Show CpuModel and CpuMaxFreq properly
Not all cpu model name have attached freq.
Most simple case is in AMD where we get something like this in /proc/cpuinfo

	model name      : AMD Ryzen 7 1800X Eight-Core Processor
	stepping        : 1
	cpu MHz         : 3594.123

The current code breaks the program and we end up with nothing in CPU Model as well as freq.
Read Model and freq in two cases to fix.