Added Double-Click to Stacer's System Tray Icon
When Stacer is closed into the system tray it has always been inconvenient (ie. requires multiple clicks) to restore Stacer's window.  One has to click the system tray to reveal the menu and then click again one of the menu choices to bring up the Stacer window  This "fix" gives the user an extra power -- to just double-click Stacer's system tray icon to bring up the last window area s/he was on.